+++ packages/lib-shared/src/api/WDONetworkingObjects.ts
@@ -77,10 +77,12 @@ export enum WDOIdentityVerificationPhase {
     documentVerificationFinal = "DOCUMENT_VERIFICATION_FINAL",
     /** OTP verification needed */
     otp = "OTP_VERIFICATION",
-    /** Completed */
-    completed = "COMPLETED",
+    /** Waiting for onboarding approval */
+    onboardingApproval = "ONBOARDING_APPROVAL",
     /** Exchanging the temporary activation for the permanent one. */
-    activationFinish = "ACTIVATION_FINISH"
+    activationFinish = "ACTIVATION_FINISH",
+    /** Completed */
+    completed = "COMPLETED"
 }
 
 /* @internal */